[DatasetComparison] moved components to component folder
