+++ metaspace/webapp/src/components/MultiChannelController/MultiChannelController.tsx
@@ -1,15 +1,15 @@
 import { defineComponent, onMounted, onUnmounted, reactive } from '@vue/composition-api'
-import FadeTransition from '../../../../components/FadeTransition'
-import { Button } from '../../../../lib/element-ui'
-import CandidateMoleculesPopover from '../../../Annotations/annotation-widgets/CandidateMoleculesPopover.vue'
-import MolecularFormula from '../../../../components/MolecularFormula'
-import VisibleIcon from '../../../../assets/inline/refactoring-ui/icon-view-visible.svg'
-import HiddenIcon from '../../../../assets/inline/refactoring-ui/icon-view-hidden.svg'
-import RangeSlider from '../../../../components/Slider/RangeSlider.vue'
-import IonIntensity from '../../../ImageViewer/IonIntensity.vue'
-import getColorScale from '../../../../lib/getColorScale'
-import { THUMB_WIDTH } from '../../../../components/Slider'
-import ChannelSelector from '../../../ImageViewer/ChannelSelector.vue'
+import FadeTransition from '../FadeTransition'
+import { Button } from '../../lib/element-ui'
+import CandidateMoleculesPopover from '../../modules/Annotations/annotation-widgets/CandidateMoleculesPopover.vue'
+import MolecularFormula from '../MolecularFormula'
+import VisibleIcon from '../../assets/inline/refactoring-ui/icon-view-visible.svg'
+import HiddenIcon from '../../assets/inline/refactoring-ui/icon-view-hidden.svg'
+import RangeSlider from '../Slider/RangeSlider.vue'
+import IonIntensity from '../../modules/ImageViewer/IonIntensity.vue'
+import getColorScale from '../../lib/getColorScale'
+import { THUMB_WIDTH } from '../Slider'
+import ChannelSelector from '../../modules/ImageViewer/ChannelSelector.vue'
 import './MultiChannelController.scss'
 
 interface MultiChannelControllerProps {
