+++ metaspace/webapp/src/components/SimpleIonImageViewer/SimpleIonImageViewer.scss
