+++ metaspace/webapp/src/components/SimpleIonImageViewer/SimpleIonImageViewer.tsx
@@ -1,17 +1,17 @@
 import { computed, defineComponent, onMounted, reactive, ref, watch } from '@vue/composition-api'
 import './SimpleIonImageViewer.scss'
-import IonImageViewer from '../../../../components/IonImageViewer'
-import fitImageToArea, { FitImageToAreaResult } from '../../../../lib/fitImageToArea'
-import { ImagePosition } from '../../../ImageViewer/ionImageState'
-import { IonImage, loadPngFromUrl, processIonImage, renderScaleBar } from '../../../../lib/ionImageRendering'
-import config from '../../../../lib/config'
-import createColormap from '../../../../lib/createColormap'
+import IonImageViewer from '../IonImageViewer'
+import fitImageToArea, { FitImageToAreaResult } from '../../lib/fitImageToArea'
+import { ImagePosition } from '../../modules/ImageViewer/ionImageState'
+import { IonImage, loadPngFromUrl, processIonImage, renderScaleBar } from '../../lib/ionImageRendering'
+import config from '../../lib/config'
+import createColormap from '../../lib/createColormap'
 import Vue from 'vue'
-import safeJsonParse from '../../../../lib/safeJsonParse'
-import ImageSaver from '../../../ImageViewer/ImageSaver.vue'
-import FadeTransition from '../../../../components/FadeTransition'
-import OpacitySettings from '../../../ImageViewer/OpacitySettings.vue'
-import { MultiChannelController } from './MultiChannelController'
+import safeJsonParse from '../../lib/safeJsonParse'
+import ImageSaver from '../../modules/ImageViewer/ImageSaver.vue'
+import FadeTransition from '../FadeTransition'
+import OpacitySettings from '../../modules/ImageViewer/OpacitySettings.vue'
+import { MultiChannelController } from '../MultiChannelController/MultiChannelController'
 import { cloneDeep, isEqual } from 'lodash'
 
 interface SimpleIonImageViewerProps {
